Fix bug in initialize_field()
numpy.random.Generator.random() only supports two dtypes: numpy.float32
and numpy.float64. This function raises an error when trying to
initialize a field in other dtypes, e.g., numpy.float16. Instead, we use
the default dtype with the generator and later create a copy with the
desired dtype using astype().

+++ projects/2023/project12_highlevel_programming/common.py
@@ -22,7 +22,9 @@ def initialize_field(NX, NY, NZ, dim_order="ZYX", mode="random", num_halo=0, arr
 
     assert isinstance(mode, str)
     if mode == "random":
-        tmp = rng.random(size=[NZ, NY - 2 * num_halo, NX - 2 * num_halo], dtype=dtype)
+        # random.Generator.random() only supports np.float32 and np.float64
+        # for that reason we create a copy with the desired dtype using astype()
+        tmp = rng.random(size=[NZ, NY - 2 * num_halo, NX - 2 * num_halo]).astype(dtype)
         # Uniformly distributed in [-1, 1)
         field[:, num_halo : NY - num_halo, num_halo : NX - num_halo] = 2 * tmp - 1  
     elif mode == "horizontal-bars":
